Make nameable attribute accessible
