--- baseDecorator.rb
@@ -1,27 +1,0 @@
-require_relative 'nameable'
-
-class BaseDecorator < Nameable
-    def initialize(nameable)
-        @nameable = nameable
-    end
-
-    def correct_name
-        @nameable.correct_name
-    end
-end
-
-class CapitalizeDecorator < Nameable
-    def correct_name
-        @nameable.correct_name.capitalize
-    end
-end
-
-class TrimmerDecorator < BaseDecorator
-    def correct_name
-        if @nameable.correct_name.length > 10
-            super[0..9]
-        else
-            super
-        end
-    end
-end
